add dynamic header and footer
