+++ src/js/main.js
@@ -1,10 +1,11 @@
+import { loadHeaderFooter } from "./utils.mjs";
 import ProductData from "./ProductData.mjs";
 import ProductList from "./ProductList.mjs";
-import { renderCartLength } from "./utils.mjs";
+
+loadHeaderFooter();
 
 const dataSource = new ProductData("tents");
 const element = document.querySelector(".product-list");
 const listing = new ProductList("Tents", dataSource, element);
-listing.init();
 
-renderCartLength();
+listing.init();
