+++ src/js/product.js
@@ -1,6 +1,8 @@
 import ProductData from "./ProductData.mjs";
 import ProductDetails from "./ProductDetails.mjs";
-import { getParams } from "./utils.mjs";
+import { getParams, loadHeaderFooter } from "./utils.mjs";
+
+loadHeaderFooter();
 
 const productId = getParams("product");
 const dataSource = new ProductData("tents");
