+++ src/js/utils.mjs
@@ -42,6 +42,31 @@ export function renderListWithTemplate(
   parentElement.insertAdjacentHTML(position, elements.join(""));
 }
 
+export function renderWithTemplate(template, parentElement, data, callback) {
+  parentElement.insertAdjacentHTML("afterbegin", template);
+  if (callback) {
+    callback(data);
+  }
+}
+
+async function loadTemplate(path) {
+  const res = await fetch(path);
+  const template = await res.text();
+  return template;
+}
+
+export async function loadHeaderFooter() {
+  const headerTemplate = await loadTemplate("../partials/header.html");
+  const headerElement = document.querySelector("#main-header");
+  const footerTemplate = await loadTemplate("../partials/footer.html");
+  const footerElement = document.querySelector("#main-footer");
+
+  renderWithTemplate(headerTemplate, headerElement);
+  renderWithTemplate(footerTemplate, footerElement);
+
+  renderCartLength();
+}
+
 export const counterCartLength = (cartItems) =>
   cartItems?.reduce(
     (cartLength, currentItem) => cartLength + currentItem.Quantity,
